finished API call to the movie db. did some styling with angular material

+++ server/public/scripts/controllers/genre.controller.js
@@ -4,6 +4,12 @@ app.controller('GenreController', ['MovieService', '$mdDialog', '$http', functio
     self.genre = {
         list: []
     };
+    self.getCountGenre = {
+        count: []
+    };
+    self.getImage = {
+        pics: []
+    };
 
     self.addGenre = function (genreAdd) {
         console.log('addGenre', genreAdd);
@@ -26,23 +32,43 @@ app.controller('GenreController', ['MovieService', '$mdDialog', '$http', functio
             }).then((response) => {
                 console.log('response', response);
                 self.genre.list = response.data;
+                self.getCount();
             })
             .catch((error) => {
                 console.log('error making genre GET', error);
                 alert('Something went wrong! Check the server.');
             });
     }
-    self.deleteGenre = function (genre_id) {
+    self.getCount = function () {
         $http({
-            method: 'DELETE',
-            url: `/genre/${genre_id}`
-        }).then((response) => {
-            self.getGenre();
-            alert('Success!');
-        }).catch((error) => {
-            console.log('error making genre DELETE request', error);
-            alert('Something went wrong! Check the server.');
-        });
+                method: 'GET',
+                url: '/count'
+            }).then((response) => {
+                console.log('response', response);
+                self.getCountGenre.count = response.data;
+            })
+            .catch((error) => {
+                console.log('error making genre GET', error);
+                alert('Something went wrong! Check the server.');
+            });
+    }
+
+    self.deleteGenre = function (genre) {
+        if (genre.count == 0) {
+            $http({
+                method: 'DELETE',
+                url: `/genre/${genre.id}`
+            }).then((response) => {
+                self.getGenre();
+                alert('Success!');
+            }).catch((error) => {
+                console.log('error making genre DELETE request', error);
+                alert('Something went wrong! Check the server.');
+            });
+        }
+        else {
+            alert('Genre Not Empty!!');
+        }
     }
     self.getGenre();
 }]);
